refactor:tweak the file structure.
