+++ packages/ui/src/components/DataStory/modals/nodeSettingsModal/tabs/outputSchemas/index.tsx
@@ -1,0 +1,1 @@
+export { OutputSchemas } from './outputSchemas';
